Removed bug for capacity more than 1
The center of mass and quadrant mass were not getting summed when one more point was added. It would have just updated the center of mass and mass of the quadrant to the latest inserted point specs. I fixed that now. It will now add the mass to the existing mass of the quadrant and calculate the new COM.